v1.1 Update: Treat NaNs as equal, coerce Objects to Struct
Moves to isequaln to treat NaNs as equal so that A==B if they
are otherwise equal with internal fields set to NaN.

Attempts to coerce objects to their Struct representation to enable
a broader range of comparable items with the function. This is particularly
useful since an increasing number of internal MATLAB functions return objects
instead of structs in modern releases.

Tested in R2024a.
---